requirements.txt
pyairtable>=2.3.0
openai>=1.30.0
python-dotenv>=1.0.0
rich>=13.7.0
click>=8.1.0
flask>=3.0.0
